+++ camerax_android/android/src/main/kotlin/dev/zeekr/camerax_android/common/PermissionManager.kt
@@ -7,6 +7,8 @@ import androidx.core.content.ContextCompat
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.PluginRegistry
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
@@ -22,30 +24,32 @@ object PermissionManager : ActivityAware {
     }
 
     suspend fun requestPermissions(permissions: List<Permission>): Boolean {
-        return suspendCoroutine { continuation ->
-            try {
-                val binding = binding ?: throw NullPointerException("Activity binding is null")
-                val activity = binding.activity
-                val listener = object : PluginRegistry.RequestPermissionsResultListener {
-                    override fun onRequestPermissionsResult(
-                        requestCode: Int, permissions: Array<out String>, grantResults: IntArray
-                    ): Boolean {
-                        if (requestCode != REQUEST_CODE) {
-                            return false
+        return withContext(Dispatchers.Default) {
+            suspendCoroutine { continuation ->
+                try {
+                    val binding = binding ?: throw NullPointerException("Activity binding is null")
+                    val activity = binding.activity
+                    val listener = object : PluginRegistry.RequestPermissionsResultListener {
+                        override fun onRequestPermissionsResult(
+                            requestCode: Int, permissions: Array<out String>, grantResults: IntArray
+                        ): Boolean {
+                            if (requestCode != REQUEST_CODE) {
+                                return false
+                            }
+                            // grantResults could be empty if the permissions request with the user is interrupted
+                            // https://developer.android.com/reference/android/app/Activity#onRequestPermissionsResult(int,%20java.lang.String[],%20int[])
+                            val isGranted =
+                                grantResults.isNotEmpty() && grantResults.all { it == PackageManager.PERMISSION_GRANTED }
+                            continuation.resume(isGranted)
+                            binding.removeRequestPermissionsResultListener(this)
+                            return true
                         }
-                        // grantResults could be empty if the permissions request with the user is interrupted
-                        // https://developer.android.com/reference/android/app/Activity#onRequestPermissionsResult(int,%20java.lang.String[],%20int[])
-                        val isGranted =
-                            grantResults.isNotEmpty() && grantResults.all { it == PackageManager.PERMISSION_GRANTED }
-                        continuation.resume(isGranted)
-                        binding.removeRequestPermissionsResultListener(this)
-                        return true
                     }
+                    binding.addRequestPermissionsResultListener(listener)
+                    ActivityCompat.requestPermissions(activity, permissions.values, REQUEST_CODE)
+                } catch (e: Exception) {
+                    continuation.resumeWithException(e)
                 }
-                binding.addRequestPermissionsResultListener(listener)
-                ActivityCompat.requestPermissions(activity, permissions.values, REQUEST_CODE)
-            } catch (e: Exception) {
-                continuation.resumeWithException(e)
             }
         }
     }