+++ camerax_ios/ios/camerax_ios/Sources/camerax_ios/common/PermissionManagerDelegate.swift
@@ -23,11 +23,9 @@ class PermissionManagerDelegate: PigeonApiDelegatePermissionManagerProxyApi {
 
 extension PermissionManagerPermissionApi {
     var delegate: PermissionManager.Permission {
-        switch self {
-        case .video:
-            return .video
-        case .audio:
-            return .audio
+        return switch self {
+        case .video: .video
+        case .audio: .audio
         }
     }
 }